refactor: convert values representing sizes/lengths and counters to type `ssize_t`; fix style issues

+++ 4/1-replace-word.c
@@ -13,9 +13,9 @@ int main(int argc, const char* const* argv)
     if (argc < REQUIRED_ARG_COUNT + 1)
         exit(1);
 
-    int position = atoi(argv[1]);
+    ssize_t position = atol(argv[1]);
 
-    int word_length = strlen(argv[2]);
+    ssize_t word_length = strlen(argv[2]);
     if (word_length <= 0)
         exit(2);
